feat: add event-list component
