+++ app/organization/page.tsx
@@ -3,10 +3,10 @@
 import AvatarOrganization from "@/components/organization/avatar"
 import BadgeSection from "@/components/organization/badge-section"
 import BannerOrganization from "@/components/organization/banner"
-import BlogPost from "@/components/organization/blog-post"
 import CreateEvent from "@/components/organization/create-event"
 import EventName from "@/components/organization/event-name"
 import ProfileInfo from "@/components/organization/profile-info"
+import EventList from "@/components/organization/event-list"
 
 const eventData = [
   {
@@ -83,27 +83,6 @@ const badgeSponsorData = {
   ],
 }
 
-const eventHostedData = [
-  {
-    imageSrc: "/organization/team.png",
-    tag: "Most popular",
-    title: "Many desktop publishing packages and web page",
-    description:
-      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Maecenas quis eros sed risus sollicitudin fringilla dictum in metus.",
-    avatarSrc: "/organization/avatar.png",
-    organizationName: "WebXGuild",
-  },
-  {
-    imageSrc: "/organization/team.png",
-    tag: "Most popular",
-    title: "Many desktop publishing packages and web page",
-    description:
-      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Maecenas quis eros sed risus sollicitudin fringilla dictum in metus.",
-    avatarSrc: "/organization/avatar.png",
-    organizationName: "WebXGuild",
-  },
-]
-
 const OrganizationPage = () => {
   return (
     <div className="container mb-20">
@@ -156,19 +135,7 @@ const OrganizationPage = () => {
         <div className="border-b-1 mt-16 w-full border" />
         <div>
           <h1 className="mt-12 text-2xl">Events Hosted</h1>
-          <div className="mt-8 flex justify-evenly space-x-5">
-            {eventHostedData.map((event, index) => (
-              <BlogPost
-                key={index}
-                imageSrc={event.imageSrc}
-                tag={event.tag}
-                title={event.title}
-                description={event.description}
-                avatarSrc={event.avatarSrc}
-                orgName={event.organizationName}
-              />
-            ))}
-          </div>
+          <EventList />
         </div>
       </div>
     </div>
